refactor(provider): extract shared stream idle-timeout modules
The streaming idle-timeout budget and its reasoning-effort scaling were
inlined in provider/mod.rs and openai_stream_runtime.rs, two files already
at their code-size budget, which made the shared contract between the
HTTPS/SSE and websocket transports hard to see and test.

- move the provider-agnostic budget helpers into
  jcode-base/src/provider/stream_timeout.rs, re-exported unchanged
- move the OpenAI Responses timeout/reasoning-payload helpers into
  openai_stream_timeout.rs, including the `summary: "auto"` payload builder
  that keeps high-effort turns from streaming silently
- add tests covering effort normalization, the multiplier ceiling, unknown
  future efforts, and that both transports agree on effort ordering
- ratchet the code-size budget down for the files that shrank
