+++ crates/jcode-provider-openai-runtime/src/openai_tests/transport_runtime.rs
@@ -560,49 +560,3 @@ async fn persistent_ws_does_not_reuse_response_cancelled_before_completion() {
     );
     server.await.expect("test websocket server");
 }
-
-use crate::openai_stream_runtime::{effective_ws_completion_timeout_secs, request_reasoning_effort};
-
-#[test]
-fn ws_completion_timeout_scales_with_request_reasoning_effort() {
-    let base = effective_ws_completion_timeout_secs(&serde_json::json!({"model": "gpt-5.6"}));
-    assert!(base > 0);
-
-    // A request without reasoning keeps the base budget.
-    assert_eq!(
-        effective_ws_completion_timeout_secs(
-            &serde_json::json!({"reasoning": {"effort": "low", "summary": "auto"}})
-        ),
-        base
-    );
-
-    // Max effort can think silently far longer than an ordinary turn, so the
-    // websocket completion budget must grow with it instead of killing the
-    // stream mid-thought.
-    let max_effort = effective_ws_completion_timeout_secs(
-        &serde_json::json!({"reasoning": {"effort": "max", "summary": "auto"}}),
-    );
-    assert!(
-        max_effort > base,
-        "max effort budget {max_effort}s should exceed base {base}s"
-    );
-    assert!(
-        effective_ws_completion_timeout_secs(
-            &serde_json::json!({"reasoning": {"effort": "xhigh"}})
-        ) > base
-    );
-}
-
-#[test]
-fn request_reasoning_effort_reads_the_responses_payload_shape() {
-    assert_eq!(
-        request_reasoning_effort(&serde_json::json!({"reasoning": {"effort": "high"}})),
-        Some("high")
-    );
-    assert_eq!(request_reasoning_effort(&serde_json::json!({})), None);
-    // Malformed shapes must not panic or be mistaken for an effort.
-    assert_eq!(
-        request_reasoning_effort(&serde_json::json!({"reasoning": "high"})),
-        None
-    );
-}
