+++ scripts/code_size_budget.json
@@ -27,7 +27,7 @@
     "crates/jcode-base/src/memory.rs": 2065,
     "crates/jcode-base/src/memory_agent.rs": 1901,
     "crates/jcode-base/src/provider/catalog_routes.rs": 1438,
-    "crates/jcode-base/src/provider/mod.rs": 2742,
+    "crates/jcode-base/src/provider/mod.rs": 2737,
     "crates/jcode-base/src/session.rs": 1621,
     "crates/jcode-base/src/sidecar.rs": 1319,
     "crates/jcode-base/src/skill.rs": 1431,
@@ -54,8 +54,8 @@
     "crates/jcode-provider-doctor/src/live_provider_probes.rs": 2029,
     "crates/jcode-provider-doctor/src/provider_e2e.rs": 2713,
     "crates/jcode-provider-openai-runtime/src/lib.rs": 1371,
-    "crates/jcode-provider-openai-runtime/src/openai_stream_runtime.rs": 1651,
-    "crates/jcode-provider-openrouter-runtime/src/lib.rs": 2756,
+    "crates/jcode-provider-openai-runtime/src/openai_stream_runtime.rs": 1650,
+    "crates/jcode-provider-openrouter-runtime/src/lib.rs": 2745,
     "crates/jcode-render-core/src/math.rs": 1250,
     "crates/jcode-setup-hints/src/lib.rs": 2589,
     "crates/jcode-telemetry-core/src/lib.rs": 2101,
